Move role_id specific api routes to correct file
